requirements.txt
black
celery
dj-database-url
dj-static
Django
django-admin-rangefilter
django-celery-beat
django-dbbackup
django-debug-toolbar
django-extensions
django-filter
django-logentry-admin
djangorestframework-simplejwt
djangorestframework
static3
gunicorn
jsondate
psycopg2-binary
python-decouple
pytz
raven
redis
